Refactor and move toggle to core.
