--- src/components/Form/Form.tsx
@@ -1,1 +1,0 @@
-export { Toggle, ToggleProps, ToggleInput, ToggleInputProps } from './Toggle';
