+++ src/components/index.ts
@@ -1,11 +1,5 @@
 export { Breadcrumb, BreadcrumbProps } from './Breadcrumb/Breadcrumb';
 export { Block, BlockProps } from './Block/Block';
-export {
-  Toggle,
-  ToggleProps,
-  ToggleInputProps,
-  ToggleInput,
-} from './Form/Toggle';
 export { Heading, HeadingProps } from './Heading/Heading';
 export { Link, LinkProps } from './Link/Link';
 export { LinkExternal, LinkExternalProps } from './Link/LinkExternal';
