+++ src/core/Form/Form.tsx
@@ -1,9 +1,4 @@
-export {
-  Toggle,
-  ToggleProps,
-  ToggleInput,
-  ToggleInputProps,
-} from './Toggle/Toggle';
+export { ToggleProps, ToggleInput, ToggleButton } from './Toggle';
 export { TextInput, TextInputProps } from './TextInput/TextInput';
 export { SearchInput, SearchInputProps } from './SearchInput/SearchInput';
 export { Checkbox, CheckboxProps } from './Checkbox/Checkbox';
